Clean up code
Simplify and refactor. Place `Article` type in `ArticleCard` as it makes most sense there. Remove code duplication by passing an `Article` to `ArticleCard` instead of title, imgUrl and readMoreUrl as props.

Remove unnecessary code in `Blog`. Push the article object to the list without the image, and update it with the image if the image exists (otherwise articles that don't have an image won't be loaded).

Simplify the `BlogList` component by checking the articles array length inside the `ArticleColumn` as opposed to outside.