Add "Nothing to show" blog card
In the case that there are no blog posts to show, display a "nothing to show" card.

Add loading logic - when the request is happening, the "loaded" boolean is "false", when the loading finishes, it is set to "true". This way, there's a difference between no blogs being displayed because it's still loading, and when there are no blogs displayed when there are no blogs written by the author.

If it's still loading, display skeletons, if its loaded and it's empty, show "Nothing to show" card. If it's loaded and contains blogs, show those blogs.