Clean up pages
Clean up Home page - remove InstaFeed (it's in About page now). Add a 'start now' button to encourage the user to make the purchase.

Tidy up the About page - make it more compact, give width of 400px at xs and up - that way it doesn't look weird on smaller devices when the width is smaller than sm (it used to be set to 100% and that just looks too goofy at 599px)